[BUG] Fix issue with Travis commit sha
When using "TRAVIS_COMMIT" we were expecting this to be the SHA of the latest commit on the pull request.

It turns out that Travis first merges the base branch into the branch under test, which means the latest commit (the one references by TRAVIS_COMMIT) is a merge commit, and not the one that we want to test.

To work around this, this commit instead uses "TRAVIS_PULL_REQUEST_SHA" to get the commit SHA.